little fix on hp values, updated texture files

+++ src/main/java/com/bluexin/saoui/SAOIngameGUI.java
@@ -447,12 +447,12 @@ public class SAOIngameGUI extends GuiIngame {
         if (!SAOOption.REMOVE_HPXP.value) {
             String absorb = SAOOption.ALT_ABSORB_POS.value? "":" ";
             if (mc.thePlayer.getAbsorptionAmount() > 0) {
-                absorb += "(+" + Math.ceil(mc.thePlayer.getAbsorptionAmount());
+                absorb += "(+" + (int) Math.ceil(mc.thePlayer.getAbsorptionAmount());
                 absorb += ')';
                 absorb += SAOOption.ALT_ABSORB_POS.value? ' ':"";
             }
 
-            final String healthStr = String.valueOf((SAOOption.ALT_ABSORB_POS.value? absorb:"") + Math.ceil(SAOMod.getHealth(mc, mc.thePlayer, time))) + (SAOOption.ALT_ABSORB_POS.value? "":absorb) + " / " + String.valueOf(Math.ceil(SAOMod.getMaxHealth(mc.thePlayer)));
+            final String healthStr = String.valueOf((SAOOption.ALT_ABSORB_POS.value? absorb:"") + (int) Math.ceil(SAOMod.getHealth(mc, mc.thePlayer, time))) + (SAOOption.ALT_ABSORB_POS.value? "":absorb) + " / " + String.valueOf((int) Math.ceil(SAOMod.getMaxHealth(mc.thePlayer)));
             final int healthStrWidth = fontRenderer.getStringWidth(healthStr);
 
             final int absStart = healthStr.indexOf('(');
